refactor(parser): add parse previews methods

+++ src/main/java/com/github/lindenhoney/repository/SongRepository.java
@@ -3,6 +3,6 @@ package com.github.lindenhoney.repository;
 import com.github.lindenhoney.domain.Song;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 
-public interface SongRepository extends ReactiveMongoRepository<Song, Integer> {
+public interface SongRepository extends ReactiveMongoRepository<Song, Long> {
 
 }
